@@ -29,6 +29,7 @@ import (
 
 type Bottlerocket struct {
 	Options
+	EnableDefaultMountPaths bool
 }
 
 // nolint:gocyclo
@@ -95,8 +96,16 @@ func (b Bottlerocket) Script() (string, error) {
 		if s.Settings.BootstrapCommands == nil {
 			s.Settings.BootstrapCommands = map[string]BootstrapCommand{}
 		}
+		// Use appropriate bind command based on configuration
+		var bindCmd []string
+		if b.EnableDefaultMountPaths {
+			bindCmd = []string{"apiclient", "ephemeral-storage", "bind"}
+		} else {
+			bindCmd = []string{"apiclient", "ephemeral-storage", "bind", "--dirs", "/var/lib/containerd", "/var/lib/kubelet", "/var/log/pods"}
+		}
+
 		s.Settings.BootstrapCommands["000-mount-instance-storage"] = BootstrapCommand{
-			Commands:  [][]string{{"apiclient", "ephemeral-storage", "init"}, {"apiclient", "ephemeral-storage", "bind", "--dirs", "/var/lib/containerd", "/var/lib/kubelet", "/var/log/pods"}},
+			Commands:  [][]string{{"apiclient", "ephemeral-storage", "init"}, bindCmd},
 			Essential: true,
 			Mode:      BootstrapCommandModeAlways,
 		}
