@@ -80,6 +80,7 @@ var (
 		"m7a.medium":           390,
 		"m7i-flex.large":       390,
 		"r7a.medium":           390,
+		"c8i-flex.large":       468,
 		"m8i-flex.large":       468,
 		"r8i-flex.large":       468,
 		"a1.medium":            500,
@@ -99,6 +100,7 @@ var (
 		"c8gd.medium":          520,
 		"m7g.medium":           520,
 		"m7gd.medium":          520,
+		"m8a.medium":           520,
 		"m8g.medium":           520,
 		"m8gd.medium":          520,
 		"r7g.medium":           520,
@@ -155,8 +157,11 @@ var (
 		"c7gd.large":           937,
 		"c8g.large":            937,
 		"c8gd.large":           937,
+		"c8i-flex.xlarge":      937,
+		"c8i.large":            937,
 		"m7g.large":            937,
 		"m7gd.large":           937,
+		"m8a.large":            937,
 		"m8g.large":            937,
 		"m8gd.large":           937,
 		"m8i-flex.xlarge":      937,
@@ -222,8 +227,11 @@ var (
 		"c6gn.medium":          1600,
 		"c8g.xlarge":           1875,
 		"c8gd.xlarge":          1875,
+		"c8i-flex.2xlarge":     1875,
+		"c8i.xlarge":           1875,
 		"i4g.xlarge":           1875,
 		"i4i.xlarge":           1875,
+		"m8a.xlarge":           1875,
 		"m8g.xlarge":           1875,
 		"m8gd.xlarge":          1875,
 		"m8i-flex.2xlarge":     1875,
@@ -246,6 +254,7 @@ var (
 		"t4g.2xlarge":          2048,
 		"i7ie.large":           2083,
 		"inf2.xlarge":          2083,
+		"r8gb.medium":          2083,
 		"i3en.large":           2100,
 		"i8ge.large":           2100,
 		"m5dn.large":           2100,
@@ -319,13 +328,17 @@ var (
 		"r7a.2xlarge":          3125,
 		"r7i.2xlarge":          3125,
 		"r7iz.2xlarge":         3125,
+		"r8gn.medium":          3125,
 		"trn1.2xlarge":         3125,
 		"c7g.2xlarge":          3750,
 		"c7gd.2xlarge":         3750,
 		"c8g.2xlarge":          3750,
 		"c8gd.2xlarge":         3750,
+		"c8i-flex.4xlarge":     3750,
+		"c8i.2xlarge":          3750,
 		"m7g.2xlarge":          3750,
 		"m7gd.2xlarge":         3750,
+		"m8a.2xlarge":          3750,
 		"m8g.2xlarge":          3750,
 		"m8gd.2xlarge":         3750,
 		"m8i-flex.4xlarge":     3750,
@@ -342,6 +355,7 @@ var (
 		"r5dn.xlarge":          4100,
 		"r5n.xlarge":           4100,
 		"i7ie.xlarge":          4166,
+		"r8gb.large":           4166,
 		"g4ad.2xlarge":         4167,
 		"i3en.xlarge":          4200,
 		"i8ge.xlarge":          4200,
@@ -418,16 +432,20 @@ var (
 		"r7a.4xlarge":          6250,
 		"r7i.4xlarge":          6250,
 		"r7iz.4xlarge":         6250,
+		"r8gn.large":           6250,
 		"vt1.6xlarge":          6250,
 		"c6gn.xlarge":          6300,
 		"c7g.4xlarge":          7500,
 		"c7gd.4xlarge":         7500,
 		"c8g.4xlarge":          7500,
 		"c8gd.4xlarge":         7500,
+		"c8i-flex.8xlarge":     7500,
+		"c8i.4xlarge":          7500,
 		"m5a.8xlarge":          7500,
 		"m5ad.8xlarge":         7500,
 		"m7g.4xlarge":          7500,
 		"m7gd.4xlarge":         7500,
+		"m8a.4xlarge":          7500,
 		"m8g.4xlarge":          7500,
 		"m8gd.4xlarge":         7500,
 		"m8i-flex.8xlarge":     7500,
@@ -447,6 +465,7 @@ var (
 		"r5n.2xlarge":          8125,
 		"g4ad.4xlarge":         8333,
 		"i7ie.2xlarge":         8333,
+		"r8gb.xlarge":          8333,
 		"i3en.2xlarge":         8400,
 		"i8ge.2xlarge":         8400,
 		"c7i-flex.12xlarge":    9375,
@@ -494,6 +513,7 @@ var (
 		"r5d.8xlarge":          10000,
 		"x1.16xlarge":          10000,
 		"x1e.16xlarge":         10000,
+		"c8i-flex.12xlarge":    11250,
 		"m8i-flex.12xlarge":    11250,
 		"r8i-flex.12xlarge":    11250,
 		"c5.12xlarge":          12000,
@@ -555,6 +575,7 @@ var (
 		"r7a.8xlarge":          12500,
 		"r7i.8xlarge":          12500,
 		"r7iz.8xlarge":         12500,
+		"r8gn.xlarge":          12500,
 		"x2iedn.4xlarge":       12500,
 		"x2iezn.2xlarge":       12500,
 		"c5n.4xlarge":          15000,
@@ -562,12 +583,13 @@ var (
 		"c7gd.8xlarge":         15000,
 		"c8g.8xlarge":          15000,
 		"c8gd.8xlarge":         15000,
+		"c8i-flex.16xlarge":    15000,
+		"c8i.8xlarge":          15000,
 		"g4ad.8xlarge":         15000,
-		"i7i.8xlarge":          15000,
-		"i8g.8xlarge":          15000,
 		"m5zn.3xlarge":         15000,
 		"m7g.8xlarge":          15000,
 		"m7gd.8xlarge":         15000,
+		"m8a.8xlarge":          15000,
 		"m8g.8xlarge":          15000,
 		"m8gd.8xlarge":         15000,
 		"m8i-flex.16xlarge":    15000,
@@ -584,6 +606,7 @@ var (
 		"m5n.4xlarge":          16250,
 		"r5dn.4xlarge":         16250,
 		"r5n.4xlarge":          16250,
+		"r8gb.2xlarge":         16666,
 		"inf2.8xlarge":         16667,
 		"c6a.12xlarge":         18750,
 		"c6i.12xlarge":         18750,
@@ -628,10 +651,10 @@ var (
 		"c7gd.12xlarge":        22500,
 		"c8g.12xlarge":         22500,
 		"c8gd.12xlarge":        22500,
-		"i7i.12xlarge":         22500,
-		"i8g.12xlarge":         22500,
+		"c8i.12xlarge":         22500,
 		"m7g.12xlarge":         22500,
 		"m7gd.12xlarge":        22500,
+		"m8a.12xlarge":         22500,
 		"m8g.12xlarge":         22500,
 		"m8gd.12xlarge":        22500,
 		"m8i.12xlarge":         22500,
@@ -678,7 +701,9 @@ var (
 		"i3.16xlarge":          25000,
 		"i3.metal":             25000,
 		"i3en.6xlarge":         25000,
+		"i7i.8xlarge":          25000,
 		"i7ie.12xlarge":        25000,
+		"i8g.8xlarge":          25000,
 		"im4gn.4xlarge":        25000,
 		"inf1.6xlarge":         25000,
 		"is4gen.4xlarge":       25000,
@@ -726,6 +751,7 @@ var (
 		"r7iz.12xlarge":        25000,
 		"r7iz.16xlarge":        25000,
 		"r7iz.metal-16xl":      25000,
+		"r8gn.2xlarge":         25000,
 		"vt1.24xlarge":         25000,
 		"x1.32xlarge":          25000,
 		"x1e.32xlarge":         25000,
@@ -735,18 +761,20 @@ var (
 		"z1d.12xlarge":         25000,
 		"z1d.metal":            25000,
 		"i4i.12xlarge":         28120,
+		"i7i.12xlarge":         28120,
+		"i8g.12xlarge":         28120,
 		"c7g.16xlarge":         30000,
 		"c7g.metal":            30000,
 		"c7gd.16xlarge":        30000,
 		"c7gd.metal":           30000,
 		"c8g.16xlarge":         30000,
 		"c8gd.16xlarge":        30000,
-		"i7i.16xlarge":         30000,
-		"i8g.16xlarge":         30000,
+		"c8i.16xlarge":         30000,
 		"m7g.16xlarge":         30000,
 		"m7g.metal":            30000,
 		"m7gd.16xlarge":        30000,
 		"m7gd.metal":           30000,
+		"m8a.16xlarge":         30000,
 		"m8g.16xlarge":         30000,
 		"m8gd.16xlarge":        30000,
 		"m8i.16xlarge":         30000,
@@ -758,6 +786,7 @@ var (
 		"r8gd.16xlarge":        30000,
 		"r8i.16xlarge":         30000,
 		"x8g.16xlarge":         30000,
+		"r8gb.4xlarge":         33330,
 		"g6e.16xlarge":         35000,
 		"c6a.24xlarge":         37500,
 		"c6i.24xlarge":         37500,
@@ -767,7 +796,9 @@ var (
 		"c7i.metal-24xl":       37500,
 		"i4g.16xlarge":         37500,
 		"i4i.16xlarge":         37500,
+		"i7i.16xlarge":         37500,
 		"i7ie.18xlarge":        37500,
+		"i8g.16xlarge":         37500,
 		"i8ge.6xlarge":         37500,
 		"m6a.24xlarge":         37500,
 		"m6i.24xlarge":         37500,
@@ -785,9 +816,12 @@ var (
 		"c8g.metal-24xl":       40000,
 		"c8gd.24xlarge":        40000,
 		"c8gd.metal-24xl":      40000,
+		"c8i.24xlarge":         40000,
 		"d3en.6xlarge":         40000,
 		"g5.12xlarge":          40000,
 		"g6.12xlarge":          40000,
+		"m8a.24xlarge":         40000,
+		"m8a.metal-24xl":       40000,
 		"m8g.24xlarge":         40000,
 		"m8g.metal-24xl":       40000,
 		"m8gd.24xlarge":        40000,
@@ -800,10 +834,6 @@ var (
 		"r8i.24xlarge":         40000,
 		"x8g.24xlarge":         40000,
 		"x8g.metal-24xl":       40000,
-		"i7i.24xlarge":         45000,
-		"i7i.metal-24xl":       45000,
-		"i8g.24xlarge":         45000,
-		"i8g.metal-24xl":       45000,
 		"c5n.9xlarge":          50000,
 		"c6a.32xlarge":         50000,
 		"c6a.48xlarge":         50000,
@@ -825,6 +855,7 @@ var (
 		"c8gd.48xlarge":        50000,
 		"c8gd.metal-48xl":      50000,
 		"c8gn.4xlarge":         50000,
+		"c8i.32xlarge":         50000,
 		"d3en.8xlarge":         50000,
 		"g4dn.12xlarge":        50000,
 		"g4dn.16xlarge":        50000,
@@ -881,6 +912,7 @@ var (
 		"r8g.metal-48xl":       50000,
 		"r8gd.48xlarge":        50000,
 		"r8gd.metal-48xl":      50000,
+		"r8gn.4xlarge":         50000,
 		"r8i.32xlarge":         50000,
 		"u-3tb1.56xlarge":      50000,
 		"x2idn.16xlarge":       50000,
@@ -889,8 +921,15 @@ var (
 		"x8g.48xlarge":         50000,
 		"x8g.metal-48xl":       50000,
 		"i4i.24xlarge":         56250,
+		"i7i.24xlarge":         56250,
+		"i7i.metal-24xl":       56250,
+		"i8g.24xlarge":         56250,
+		"i8g.metal-24xl":       56250,
+		"r8gb.8xlarge":         66660,
 		"c6gn.12xlarge":        75000,
 		"c6in.12xlarge":        75000,
+		"c8i.48xlarge":         75000,
+		"c8i.metal-48xl":       75000,
 		"d3en.12xlarge":        75000,
 		"i4i.32xlarge":         75000,
 		"i4i.metal":            75000,
@@ -899,6 +938,8 @@ var (
 		"m5n.16xlarge":         75000,
 		"m6idn.12xlarge":       75000,
 		"m6in.12xlarge":        75000,
+		"m8a.48xlarge":         75000,
+		"m8a.metal-48xl":       75000,
 		"m8i.48xlarge":         75000,
 		"m8i.metal-48xl":       75000,
 		"r5dn.16xlarge":        75000,
@@ -910,15 +951,14 @@ var (
 		"x2idn.24xlarge":       75000,
 		"x2iedn.24xlarge":      75000,
 		"x2iezn.8xlarge":       75000,
-		"i7i.48xlarge":         90000,
-		"i7i.metal-48xl":       90000,
-		"i8g.48xlarge":         90000,
 		"c5n.18xlarge":         100000,
 		"c5n.metal":            100000,
 		"c6gn.16xlarge":        100000,
 		"c6in.16xlarge":        100000,
 		"c7gn.8xlarge":         100000,
 		"c8gn.8xlarge":         100000,
+		"c8i.96xlarge":         100000,
+		"c8i.metal-96xl":       100000,
 		"dl2q.24xlarge":        100000,
 		"f2.48xlarge":          100000,
 		"g4dn.metal":           100000,
@@ -928,8 +968,11 @@ var (
 		"hpc6a.48xlarge":       100000,
 		"i3en.24xlarge":        100000,
 		"i3en.metal":           100000,
+		"i7i.48xlarge":         100000,
+		"i7i.metal-48xl":       100000,
 		"i7ie.48xlarge":        100000,
 		"i7ie.metal-48xl":      100000,
+		"i8g.48xlarge":         100000,
 		"im4gn.16xlarge":       100000,
 		"inf1.24xlarge":        100000,
 		"inf2.48xlarge":        100000,
@@ -951,6 +994,8 @@ var (
 		"r5n.metal":            100000,
 		"r6idn.16xlarge":       100000,
 		"r6in.16xlarge":        100000,
+		"r8gb.12xlarge":        100000,
+		"r8gn.8xlarge":         100000,
 		"r8i.96xlarge":         100000,
 		"r8i.metal-96xl":       100000,
 		"u-12tb1.112xlarge":    100000,
@@ -974,6 +1019,7 @@ var (
 		"x2iezn.12xlarge":      100000,
 		"x2iezn.metal":         100000,
 		"i8ge.18xlarge":        112500,
+		"r8gb.16xlarge":        133330,
 		"c6in.24xlarge":        150000,
 		"c7gn.12xlarge":        150000,
 		"c8gn.12xlarge":        150000,
@@ -983,6 +1029,7 @@ var (
 		"m6in.24xlarge":        150000,
 		"r6idn.24xlarge":       150000,
 		"r6in.24xlarge":        150000,
+		"r8gn.12xlarge":        150000,
 		"c6in.32xlarge":        200000,
 		"c6in.metal":           200000,
 		"c7gn.16xlarge":        200000,
@@ -1001,6 +1048,9 @@ var (
 		"r6idn.metal":          200000,
 		"r6in.32xlarge":        200000,
 		"r6in.metal":           200000,
+		"r8gb.24xlarge":        200000,
+		"r8gb.metal-24xl":      200000,
+		"r8gn.16xlarge":        200000,
 		"u7in-16tb.224xlarge":  200000,
 		"u7in-24tb.224xlarge":  200000,
 		"u7in-32tb.224xlarge":  200000,
@@ -1013,12 +1063,16 @@ var (
 		"hpc7a.96xlarge":       300000,
 		"i8ge.48xlarge":        300000,
 		"i8ge.metal-48xl":      300000,
+		"r8gn.24xlarge":        300000,
+		"r8gn.metal-24xl":      300000,
 		"dl1.24xlarge":         400000,
 		"g6e.48xlarge":         400000,
 		"p4d.24xlarge":         400000,
 		"p4de.24xlarge":        400000,
 		"c8gn.48xlarge":        600000,
 		"c8gn.metal-48xl":      600000,
+		"r8gn.48xlarge":        600000,
+		"r8gn.metal-48xl":      600000,
 		"trn1.32xlarge":        800000,
 		"trn1n.32xlarge":       1600000,
 		"p5.48xlarge":          3200000,
