@@ -17,7 +17,7 @@
 // so we can get this information at runtime.
 
 // Code generated by go generate; DO NOT EDIT.
-// This file was generated at 2025-09-05T19:41:20Z
+// This file was generated at 2025-10-10T20:02:21Z
 
 // WARNING: please add @ellistarn, @bwagner5, or @jonathan-innis from aws/karpenter to reviewers
 // if you are updating this file since Karpenter is depending on this file to calculate max pods.
@@ -3303,6 +3303,306 @@ var Limits = map[string]*VPCLimits{
 	"c8gn.xlarge": {
 		Interface:               4,
 		IPv4PerInterface:        15,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         18,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 4,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"c8i-flex.12xlarge": {
+		Interface:               12,
+		IPv4PerInterface:        50,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         110,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 12,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"c8i-flex.16xlarge": {
+		Interface:               16,
+		IPv4PerInterface:        64,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         106,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 16,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"c8i-flex.2xlarge": {
+		Interface:               4,
+		IPv4PerInterface:        30,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         18,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 4,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"c8i-flex.4xlarge": {
+		Interface:               8,
+		IPv4PerInterface:        50,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         34,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 8,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"c8i-flex.8xlarge": {
+		Interface:               10,
+		IPv4PerInterface:        50,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         52,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 10,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"c8i-flex.large": {
+		Interface:               3,
+		IPv4PerInterface:        20,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         3,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 3,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"c8i-flex.xlarge": {
+		Interface:               4,
+		IPv4PerInterface:        30,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         8,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 4,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"c8i.12xlarge": {
+		Interface:               12,
+		IPv4PerInterface:        50,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         110,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 12,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"c8i.16xlarge": {
+		Interface:               16,
+		IPv4PerInterface:        64,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         106,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 16,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"c8i.24xlarge": {
+		Interface:               16,
+		IPv4PerInterface:        64,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         106,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 16,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"c8i.2xlarge": {
+		Interface:               4,
+		IPv4PerInterface:        30,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         38,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 4,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"c8i.32xlarge": {
+		Interface:               24,
+		IPv4PerInterface:        64,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         98,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 24,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"c8i.48xlarge": {
+		Interface:               24,
+		IPv4PerInterface:        64,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         98,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 24,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"c8i.4xlarge": {
+		Interface:               8,
+		IPv4PerInterface:        50,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         54,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 8,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"c8i.8xlarge": {
+		Interface:               10,
+		IPv4PerInterface:        50,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         82,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 10,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"c8i.96xlarge": {
+		Interface:               24,
+		IPv4PerInterface:        64,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         98,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 24,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"c8i.large": {
+		Interface:               3,
+		IPv4PerInterface:        20,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         9,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 3,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"c8i.metal-48xl": {
+		Interface:               24,
+		IPv4PerInterface:        64,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         98,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 24,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "",
+		IsBareMetal: true,
+	},
+	"c8i.metal-96xl": {
+		Interface:               24,
+		IPv4PerInterface:        64,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         98,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 24,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "",
+		IsBareMetal: true,
+	},
+	"c8i.xlarge": {
+		Interface:               4,
+		IPv4PerInterface:        30,
 		IsTrunkingCompatible:    true,
 		BranchInterface:         18,
 		DefaultNetworkCardIndex: 0,
@@ -8861,6 +9161,186 @@ var Limits = map[string]*VPCLimits{
 		Hypervisor:  "nitro",
 		IsBareMetal: false,
 	},
+	"m8a.12xlarge": {
+		Interface:               12,
+		IPv4PerInterface:        64,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         110,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 12,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"m8a.16xlarge": {
+		Interface:               16,
+		IPv4PerInterface:        64,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         106,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 16,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"m8a.24xlarge": {
+		Interface:               16,
+		IPv4PerInterface:        64,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         106,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 16,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"m8a.2xlarge": {
+		Interface:               4,
+		IPv4PerInterface:        40,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         38,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 4,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"m8a.48xlarge": {
+		Interface:               24,
+		IPv4PerInterface:        64,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         98,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 24,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"m8a.4xlarge": {
+		Interface:               8,
+		IPv4PerInterface:        40,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         54,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 8,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"m8a.8xlarge": {
+		Interface:               10,
+		IPv4PerInterface:        40,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         82,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 10,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"m8a.large": {
+		Interface:               3,
+		IPv4PerInterface:        20,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         9,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 3,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"m8a.medium": {
+		Interface:               2,
+		IPv4PerInterface:        4,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         4,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 2,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"m8a.metal-24xl": {
+		Interface:               16,
+		IPv4PerInterface:        64,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         106,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 16,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "",
+		IsBareMetal: true,
+	},
+	"m8a.metal-48xl": {
+		Interface:               24,
+		IPv4PerInterface:        64,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         98,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 24,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "",
+		IsBareMetal: true,
+	},
+	"m8a.xlarge": {
+		Interface:               4,
+		IPv4PerInterface:        20,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         18,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 4,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
 	"m8g.12xlarge": {
 		Interface:               8,
 		IPv4PerInterface:        30,
@@ -9520,6 +10000,36 @@ var Limits = map[string]*VPCLimits{
 		},
 		Hypervisor:  "nitro",
 		IsBareMetal: false,
+	},
+	"mac-m4.metal": {
+		Interface:               8,
+		IPv4PerInterface:        30,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         6,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 8,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "",
+		IsBareMetal: true,
+	},
+	"mac-m4pro.metal": {
+		Interface:               8,
+		IPv4PerInterface:        30,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         6,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 8,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "",
+		IsBareMetal: true,
 	},
 	"mac1.metal": {
 		Interface:               8,
@@ -13291,6 +13801,141 @@ var Limits = map[string]*VPCLimits{
 		Hypervisor:  "nitro",
 		IsBareMetal: false,
 	},
+	"r8gb.12xlarge": {
+		Interface:               12,
+		IPv4PerInterface:        30,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         50,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 12,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"r8gb.16xlarge": {
+		Interface:               16,
+		IPv4PerInterface:        50,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         106,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 16,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"r8gb.24xlarge": {
+		Interface:               24,
+		IPv4PerInterface:        50,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         98,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 24,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"r8gb.2xlarge": {
+		Interface:               4,
+		IPv4PerInterface:        15,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         38,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 4,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"r8gb.4xlarge": {
+		Interface:               8,
+		IPv4PerInterface:        30,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         54,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 8,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"r8gb.8xlarge": {
+		Interface:               10,
+		IPv4PerInterface:        30,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         52,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 10,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"r8gb.large": {
+		Interface:               3,
+		IPv4PerInterface:        10,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         9,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 3,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"r8gb.medium": {
+		Interface:               2,
+		IPv4PerInterface:        4,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         4,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 2,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"r8gb.xlarge": {
+		Interface:               4,
+		IPv4PerInterface:        15,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         18,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 4,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
 	"r8gd.12xlarge": {
 		Interface:               8,
 		IPv4PerInterface:        30,
@@ -13457,6 +14102,161 @@ var Limits = map[string]*VPCLimits{
 		IsBareMetal: true,
 	},
 	"r8gd.xlarge": {
+		Interface:               4,
+		IPv4PerInterface:        15,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         18,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 4,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"r8gn.12xlarge": {
+		Interface:               12,
+		IPv4PerInterface:        30,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         50,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 12,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"r8gn.16xlarge": {
+		Interface:               16,
+		IPv4PerInterface:        50,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         106,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 16,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"r8gn.24xlarge": {
+		Interface:               24,
+		IPv4PerInterface:        50,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         98,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 24,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"r8gn.2xlarge": {
+		Interface:               4,
+		IPv4PerInterface:        15,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         38,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 4,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"r8gn.48xlarge": {
+		Interface:               24,
+		IPv4PerInterface:        50,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         98,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 12,
+				NetworkCardIndex:         0,
+			},
+
+			{
+				MaximumNetworkInterfaces: 12,
+				NetworkCardIndex:         1,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"r8gn.4xlarge": {
+		Interface:               8,
+		IPv4PerInterface:        30,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         54,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 8,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"r8gn.8xlarge": {
+		Interface:               10,
+		IPv4PerInterface:        30,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         52,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 10,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"r8gn.large": {
+		Interface:               3,
+		IPv4PerInterface:        10,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         9,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 3,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"r8gn.medium": {
+		Interface:               2,
+		IPv4PerInterface:        4,
+		IsTrunkingCompatible:    true,
+		BranchInterface:         4,
+		DefaultNetworkCardIndex: 0,
+		NetworkCards: []NetworkCard{
+			{
+				MaximumNetworkInterfaces: 2,
+				NetworkCardIndex:         0,
+			},
+		},
+		Hypervisor:  "nitro",
+		IsBareMetal: false,
+	},
+	"r8gn.xlarge": {
 		Interface:               4,
 		IPv4PerInterface:        15,
 		IsTrunkingCompatible:    true,
